Code Splitting for ModConf
