+++ system/usr/share/mmrl/config/livebootmagisk/json/backgroundsList.json
@@ -1,0 +1,14 @@
+[
+  {
+    "name": "Default",
+    "value": " "
+  },
+  {
+    "name": "Transparent",
+    "value": "transparent"
+  },
+  {
+    "name": "Dark",
+    "value": "dark"
+  }
+]
