+++ system/usr/share/mmrl/config/livebootmagisk/json/initialConfig.json
@@ -1,0 +1,25 @@
+{
+  "background": "dark",
+  "colors": false,
+  "wordwrap": true,
+  "logcatlevels": {
+    "V": false,
+    "D": false,
+    "I": false,
+    "W": true,
+    "E": true,
+    "F": true,
+    "S": true
+  },
+  "logcatbuffers": {
+    "M": false,
+    "S": false,
+    "R": false,
+    "E": false,
+    "C": true
+  },
+  "logcatformat": "brief",
+  "dmesg": false,
+  "lines": "80",
+  "save": false
+}
