+++ system/usr/share/mmrl/config/livebootmagisk/json/logcatFormatsList.json
@@ -1,0 +1,26 @@
+[
+  {
+    "name": "Brief",
+    "value": "brief"
+  },
+  {
+    "name": "Process",
+    "value": "process"
+  },
+  {
+    "name": "Tag",
+    "value": "tag"
+  },
+  {
+    "name": "Thread",
+    "value": "thread"
+  },
+  {
+    "name": "Time",
+    "value": "time"
+  },
+  {
+    "name": "Thread time",
+    "value": "threadtime"
+  }
+]
