+++ system/usr/share/mmrl/config/livebootmagisk/json/logcatLevelsList.json
@@ -1,0 +1,30 @@
+[
+  {
+    "name": "Verbose",
+    "value": "V"
+  },
+  {
+    "name": "Debug",
+    "value": "D"
+  },
+  {
+    "name": "Info",
+    "value": "I"
+  },
+  {
+    "name": "Warning",
+    "value": "W"
+  },
+  {
+    "name": "Error",
+    "value": "E"
+  },
+  {
+    "name": "Fatal",
+    "value": "F"
+  },
+  {
+    "name": "Silent",
+    "value": "S"
+  }
+]
