+++ system/usr/share/mmrl/config/livebootmagisk/json/serviceFiles.json
@@ -1,0 +1,1 @@
+["/data/adb/service.d/0000bootlive", "/data/adb/post-fs-data.d/0000bootlive", "/data/adb/modules/livebootmagisk/0000bootlive"]
